put min,max,clamp in separate header #4425
