+++ core/unit_test/TestMathematicalFunctions.hpp
@@ -1016,200 +1016,3 @@ struct TestIsNaN {
 TEST(TEST_CATEGORY, mathematical_functions_isnan) {
   TestIsNaN<TEST_EXECSPACE>();
 }
-
-// ----------------------------------------------------------
-// test max()
-// ----------------------------------------------------------
-TEST(TEST_CATEGORY, max) {
-  namespace KE = Kokkos::Experimental;
-
-  int a = 1;
-  int b = 2;
-  EXPECT_TRUE(KE::max(a, b) == 2);
-
-  a = 3;
-  b = 1;
-  EXPECT_TRUE(KE::max(a, b) == 3);
-}
-
-template <class ViewType>
-struct StdAlgoMinMaxOpsTestMax {
-  ViewType m_view;
-
-  KOKKOS_INLINE_FUNCTION
-  void operator()(const int& ind) const {
-    namespace KE = Kokkos::Experimental;
-    auto v1      = 10.;
-    if (KE::max(v1, m_view(ind)) == 10.) {
-      m_view(ind) = 6.;
-    }
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  StdAlgoMinMaxOpsTestMax(ViewType aIn) : m_view(aIn) {}
-};
-
-TEST(TEST_CATEGORY, max_within_parfor) {
-  namespace KE = Kokkos::Experimental;
-
-  using view_t = Kokkos::View<double*>;
-  view_t a("a", 10);
-
-  StdAlgoMinMaxOpsTestMax<view_t> fnc(a);
-  Kokkos::parallel_for(a.extent(0), fnc);
-  auto a_h = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), a);
-  for (int i = 0; i < 10; ++i) {
-    EXPECT_DOUBLE_EQ(a_h(0), 6.);
-  }
-}
-
-// ----------------------------------------------------------
-// test min()
-// ----------------------------------------------------------
-TEST(TEST_CATEGORY, min) {
-  namespace KE = Kokkos::Experimental;
-
-  int a = 1;
-  int b = 2;
-  EXPECT_TRUE(KE::min(a, b) == 1);
-
-  a = 3;
-  b = 2;
-  EXPECT_TRUE(KE::min(a, b) == 2);
-}
-
-template <class ViewType>
-struct StdAlgoMinMaxOpsTestMin {
-  ViewType m_view;
-
-  KOKKOS_INLINE_FUNCTION
-  void operator()(const int& ind) const {
-    namespace KE = Kokkos::Experimental;
-    auto v1      = 10.;
-    if (KE::min(v1, m_view(ind)) == 0.) {
-      m_view(ind) = 8.;
-    }
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  StdAlgoMinMaxOpsTestMin(ViewType aIn) : m_view(aIn) {}
-};
-
-TEST(TEST_CATEGORY, min_within_parfor) {
-  namespace KE = Kokkos::Experimental;
-  using view_t = Kokkos::View<double*>;
-  view_t a("a", 10);
-
-  StdAlgoMinMaxOpsTestMin<view_t> fnc(a);
-  Kokkos::parallel_for(a.extent(0), fnc);
-  auto a_h = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), a);
-  for (int i = 0; i < 10; ++i) {
-    EXPECT_DOUBLE_EQ(a_h(0), 8.);
-  }
-}
-
-// ----------------------------------------------------------
-// test minmax()
-// ----------------------------------------------------------
-TEST(TEST_CATEGORY, minmax) {
-  namespace KE  = Kokkos::Experimental;
-  int a         = 1;
-  int b         = 2;
-  const auto& r = KE::minmax(a, b);
-  EXPECT_TRUE(r.first == 1);
-  EXPECT_TRUE(r.second == 2);
-
-  a              = 3;
-  b              = 2;
-  const auto& r2 = KE::minmax(a, b);
-  EXPECT_TRUE(r2.first == 2);
-  EXPECT_TRUE(r2.second == 3);
-}
-
-template <class ViewType>
-struct StdAlgoMinMaxOpsTestMinMax {
-  ViewType m_view;
-
-  KOKKOS_INLINE_FUNCTION
-  void operator()(const int& ind) const {
-    namespace KE  = Kokkos::Experimental;
-    auto v1       = 7.;
-    const auto& r = KE::minmax(v1, m_view(ind));
-    m_view(ind)   = (double)(r.first - r.second);
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  StdAlgoMinMaxOpsTestMinMax(ViewType aIn) : m_view(aIn) {}
-};
-
-TEST(TEST_CATEGORY, minmax_within_parfor) {
-  namespace KE = Kokkos::Experimental;
-  using view_t = Kokkos::View<double*>;
-  view_t a("a", 10);
-
-  StdAlgoMinMaxOpsTestMinMax<view_t> fnc(a);
-  Kokkos::parallel_for(a.extent(0), fnc);
-  auto a_h = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), a);
-  for (int i = 0; i < 10; ++i) {
-    EXPECT_DOUBLE_EQ(a_h(0), -7.);
-  }
-}
-
-// ----------------------------------------------------------
-// test clamp()
-// ----------------------------------------------------------
-TEST(TEST_CATEGORY, clamp) {
-  namespace KE = Kokkos::Experimental;
-
-  int a         = 1;
-  int b         = 2;
-  int c         = 19;
-  const auto& r = KE::clamp(a, b, c);
-  EXPECT_TRUE(&r == &b);
-  EXPECT_TRUE(r == b);
-
-  a              = 5;
-  b              = -2;
-  c              = 3;
-  const auto& r2 = KE::clamp(a, b, c);
-  EXPECT_TRUE(&r2 == &c);
-  EXPECT_TRUE(r2 == c);
-
-  a              = 5;
-  b              = -2;
-  c              = 7;
-  const auto& r3 = KE::clamp(a, b, c);
-  EXPECT_TRUE(&r3 == &a);
-  EXPECT_TRUE(r3 == a);
-}
-
-template <class ViewType>
-struct StdAlgoMinMaxOpsTestClamp {
-  ViewType m_view;
-
-  KOKKOS_INLINE_FUNCTION
-  void operator()(const int& ind) const {
-    namespace KE  = Kokkos::Experimental;
-    m_view(ind)   = 10.;
-    const auto b  = -2.;
-    const auto c  = 3.;
-    const auto& r = KE::clamp(m_view(ind), b, c);
-    m_view(ind)   = (double)(r);
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  StdAlgoMinMaxOpsTestClamp(ViewType aIn) : m_view(aIn) {}
-};
-
-TEST(TEST_CATEGORY, clamp_within_parfor) {
-  namespace KE = Kokkos::Experimental;
-  using view_t = Kokkos::View<double*>;
-  view_t a("a", 10);
-
-  StdAlgoMinMaxOpsTestClamp<view_t> fnc(a);
-  Kokkos::parallel_for(a.extent(0), fnc);
-  auto a_h = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), a);
-  for (std::size_t i = 0; i < a.extent(0); ++i) {
-    EXPECT_DOUBLE_EQ(a_h(0), 3.);
-  }
-}
